Refactor datamodule
