+++ pvnet/data/base.py
@@ -1,15 +1,17 @@
 """ Data module for pytorch lightning """
+from datetime import datetime
 
 import torch
+from lightning.pytorch import LightningDataModule
 from ocf_datapipes.batch import stack_np_examples_into_batch
 from ocf_datapipes.training.pvnet import pvnet_datapipe
+from torch.utils.data import DataLoader
 from torch.utils.data.datapipes.iter import FileLister
 
 from pvnet.data.utils import batch_to_tensor
-from pvnet.data.base import BaseDataModule
 
 
-class DataModule(BaseDataModule):
+class BaseDataModule(LightningDataModule):
     """Datamodule for training pvnet and using pvnet pipeline in `ocf_datapipes`."""
 
     def __init__(
@@ -40,8 +42,41 @@ class DataModule(BaseDataModule):
                 `configuration` or 'train/val/test_period'.
 
         """
-        super().__init__(configuration=configuration, batch_size=batch_size, num_workers=num_workers, prefetch_factor=prefetch_factor, train_period=train_period, val_period=val_period, test_period=test_period, batch_dir=batch_dir)
+        super().__init__()
+        self.configuration = configuration
+        self.batch_size = batch_size
+        self.batch_dir = batch_dir
 
+        if not ((batch_dir is not None) ^ (configuration is not None)):
+            raise ValueError("Exactly one of `batch_dir` or `configuration` must be set.")
+
+        if batch_dir is not None:
+            if any([period != [None, None] for period in [train_period, val_period, test_period]]):
+                raise ValueError("Cannot set `(train/val/test)_period` with presaved batches")
+
+        self.train_period = [
+            None if d is None else datetime.strptime(d, "%Y-%m-%d") for d in train_period
+        ]
+        self.val_period = [
+            None if d is None else datetime.strptime(d, "%Y-%m-%d") for d in val_period
+        ]
+        self.test_period = [
+            None if d is None else datetime.strptime(d, "%Y-%m-%d") for d in test_period
+        ]
+
+        self._common_dataloader_kwargs = dict(
+            batch_size=None,  # batched in datapipe step
+            sampler=None,
+            batch_sampler=None,
+            num_workers=num_workers,
+            collate_fn=None,
+            pin_memory=False,
+            drop_last=False,
+            timeout=0,
+            worker_init_fn=None,
+            prefetch_factor=prefetch_factor,
+            persistent_workers=False,
+        )
 
     def _get_datapipe(self, start_time, end_time):
         data_pipeline = pvnet_datapipe(
@@ -83,3 +118,26 @@ class DataModule(BaseDataModule):
 
         return data_pipeline
 
+    def train_dataloader(self):
+        """Construct train dataloader"""
+        if self.batch_dir is not None:
+            datapipe = self._get_premade_batches_datapipe("train", shuffle=True)
+        else:
+            datapipe = self._get_datapipe(*self.train_period)
+        return DataLoader(datapipe, shuffle=True, **self._common_dataloader_kwargs)
+
+    def val_dataloader(self):
+        """Construct val dataloader"""
+        if self.batch_dir is not None:
+            datapipe = self._get_premade_batches_datapipe("val")
+        else:
+            datapipe = self._get_datapipe(*self.val_period)
+        return DataLoader(datapipe, shuffle=False, **self._common_dataloader_kwargs)
+
+    def test_dataloader(self):
+        """Construct test dataloader"""
+        if self.batch_dir is not None:
+            datapipe = self._get_premade_batches_datapipe("test")
+        else:
+            datapipe = self._get_datapipe(*self.test_period)
+        return DataLoader(datapipe, shuffle=False, **self._common_dataloader_kwargs)
